# backend/contracts/api.yaml
openapi: 3.0.1
info:
  title: Dena Analytics API
  version: 1.0.0
servers:
- url: http://host.docker.internal:3000
paths:
...
  /api/dashboards/{id}:
    get:
      operationId: getDashboardById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          example: 1
        examples:
          dashboard1:
            value: 1
          dashboardNotFound:
            value: 999
      responses:
        '200':
          description: Dashboard
          content:
            application/json:
              schema:
                type: object
                required:
                - id
                - name
                - category
                - powerBiUrl
                - lastUpdated
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  category:
                    type: string
                  powerBiUrl:
                    type: string
                    format: uri
                  lastUpdated:
                    type: string
                    format: date-time
              examples:
                dashboard1:
                  value:
                    id: 1
                    name: Census & Population Demographics
                    category: census
                    powerBiUrl: https://example.com
                    lastUpdated: '2026-06-08T10:00:00Z'
        '404':
          description: Dashboard not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              examples:
                dashboardNotFound:
                  value:
                    error: Dashboard with ID 999 not found
        '400':
          description: Bad Request
          headers:
            X-Powered-By:
              required: true
              schema:
                type: string
            Connection:
              required: true
              schema:
                type: string
            Keep-Alive:
              required: true
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Api_Dashboards_True_GET_400_application_json_ResponseBody'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Details:
      required:
      - field
      - message
      type: object
      properties:
        field:
          type: string
        message:
          type: string
...
    Api_Dashboards_True_GET_400_application_json_ResponseBody:
      required:
      - details
      - error
      type: object
      properties:
        error:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/Details'
...